Address review feedback: scope to CFB8 only, move tests to module
