+++ patches/boringssl/expose_aes-cfb8.patch
@@ -1,6 +1,6 @@
 --- a/crypto/cipher/get_cipher.cc
 +++ b/crypto/cipher/get_cipher.cc
-@@ -32,17 +32,21 @@
+@@ -32,17 +32,20 @@
  } kCiphers[] = {
      {NID_aes_128_cbc, "aes-128-cbc", EVP_aes_128_cbc},
      {NID_aes_128_cfb128, "aes-128-cfb", EVP_aes_128_cfb128},
@@ -10,7 +10,6 @@
      {NID_aes_128_gcm, "aes-128-gcm", EVP_aes_128_gcm},
      {NID_aes_128_ofb128, "aes-128-ofb", EVP_aes_128_ofb},
      {NID_aes_192_cbc, "aes-192-cbc", EVP_aes_192_cbc},
-+    {NID_aes_192_cfb128, "aes-192-cfb", EVP_aes_192_cfb128},
 +    {NID_aes_192_cfb8, "aes-192-cfb8", EVP_aes_192_cfb8},
      {NID_aes_192_ctr, "aes-192-ctr", EVP_aes_192_ctr},
      {NID_aes_192_ecb, "aes-192-ecb", EVP_aes_192_ecb},
@@ -109,7 +108,7 @@
 +const EVP_CIPHER *EVP_aes_256_cfb8() { return &aes_256_cfb8; }
 --- a/decrepit/evp/evp_do_all.cc
 +++ b/decrepit/evp/evp_do_all.cc
-@@ -23,17 +23,21 @@
+@@ -23,17 +23,20 @@
    // Cipher lookups are case-insensitive, so uppercase names are not needed.
    callback(EVP_aes_128_cbc(), "aes-128-cbc", nullptr, arg);
    callback(EVP_aes_128_cfb128(), "aes-128-cfb", nullptr, arg);
@@ -119,7 +118,6 @@
    callback(EVP_aes_128_gcm(), "aes-128-gcm", nullptr, arg);
    callback(EVP_aes_128_ofb(), "aes-128-ofb", nullptr, arg);
    callback(EVP_aes_192_cbc(), "aes-192-cbc", nullptr, arg);
-+  callback(EVP_aes_192_cfb128(), "aes-192-cfb", nullptr, arg);
 +  callback(EVP_aes_192_cfb8(), "aes-192-cfb8", nullptr, arg);
    callback(EVP_aes_192_ctr(), "aes-192-ctr", nullptr, arg);
    callback(EVP_aes_192_ecb(), "aes-192-ecb", nullptr, arg);
